+++ source/standalone/workflows/skrl/config.py
@@ -81,5 +81,7 @@ def convert_skrl_cfg(cfg):
                 elif key in ["rewards_shaper_scale"]:
                     d["rewards_shaper"] = reward_shaper_function(value)
 
+    # parse agent configuration and convert to classes
     update_dict(cfg)
+    # return the updated configuration
     return cfg